add types/size coefficients and types.memory

+++ go/virtualization/vm.GetRam.go
@@ -2,6 +2,6 @@ package virtualization
 
 func (vm *VM) GetRam() uint {
 
-	return vm.ram
+	return uint(vm.ram)
 
 }
